Makes signatures `bigint`.
This started as a bugfix since `v` in Ethereum can actually be an arbitrarily large number due to EIP-155, and this was encoding it as 1 byte.  However, I realized that the JSON-RPC protocol _actually_ considers all 3 components of the signature to be `Quantity` (e.g., number) and since working with `bigint` is much easier than working with a byte array, I decided to just convert them all to `bigint`s.

I also swapped teh order from `v,r,s` to `r,s,v` since that aligns better both alphabetically, and how they are calculated in the `secp256k1` signing process.

Finally, renamed `Signature` to `EncodedSignature` since an actual Signature is 3 component numbers while an EncodedSignature is a byte array in a specific encoding format.  This will hopefully reduce confusion for library users (like me!).