bigint All The Things!
Switches to using `bigint` for all fixed-width byte fields.  The developer UX for `bigint` is significantly better than `Uint8Array` or even `Bytes` because you can do things like `0xADD4E55n` rather than having to start with a string and then convert it to a byte array.

Removes all of the generic stuff.  Now that we are using primitive types for everything, we can drop the generics and instead just have everything use a primitive.  The one "non-primitive" is `Uint8Array`, which is more widely available than `bigint`.