Changes all JSON-RPC properties to bigints.
While for many of these properties `number` is big enough, they all **MUST** be integers (whole numbers, no fractional component).  Since the range of allowable/reasonable values is smaller than a JS number's integer range, we don't gain anything from using `number` for any of these, and we lose guarantees about no-fractional components.  Also, this brings consistency to the interface in that you'll be working with `bigint` almost exclusively now.

The JSON-RPC `id` field has been left as a `number` since I believe that technically `1.234` is a valid ID, and `1.234n` is not.