Adds support for OpenEthereum PoA networks.
They don't include `nonce` or `mixHash` for blocks.  Technically, we should include the seal array, but I don't know what is in it and instantseal networks just have an empty array so I can ignore it for my needs.  If someone can figure out what exactly the seal array is, it can be added here as an optional block property.