Fixes incompatibility with Geth and gasPrice on eth_call.
Geth will consider gasLimit and gasPrice when executing an `eth_call` if present, and if `gasLimit` is missing it will default to 125,000,000 gas.  This means if you specify a `gasPrice` and no `gasLimit` you'll end up with a failed transaction unless you have enough ETH in the account to cover.  This change makes both `gasLimit` and `gasPrice` optional on off-chain transactions (while still requried in on-chain transactions).

* switches NPM publishing over to GitHub actions.
* updates dev dependencies
* adds declaration map support